+++ skyVegasTests/utilities/index.ts
@@ -1,3 +1,7 @@
 export { launchGame } from './launchGameHelper';
 export { validateConsoleMessages } from './validateConsoleMessagesHelper';
 export { login } from './loginHelper';
+export { startEventListener } from './startEventListenerHelper';
+export { messageExists } from './messageExistsHelper';
+export { makeSpin } from "./makeSpinHelper";
+export { makeDeposit } from "./makeDepositHelper";
