+++ skyVegasTests/utilities/messageExistsHelper.ts
@@ -1,0 +1,1 @@
+export const messageExists = (consoleMessages: string[], expectedMessage: string):boolean => consoleMessages.some((msg) => msg.includes(expectedMessage));
