--- skyVegasTests/utilities/saferGamblingHelper.ts
@@ -1,6 +1,0 @@
-import { type Page } from "@playwright/test";
-
-export let saferGamblingPage = async (sgPage:Page) => {
-    const sgUrl = process.env.SGURL!;
-    await sgPage.goto(sgUrl);
-}
