+++ skyVegasTests/utilities/validateConsoleMessagesHelper.ts
@@ -1,18 +1,29 @@
 import { expect } from "@playwright/test";
-export const messageExists = (consoleMessages: string[], expectedMessage: string):boolean => consoleMessages.some((msg) => msg.includes(expectedMessage));
+import { messageExists } from "../utilities";
 
 export const validateConsoleMessages = async (
   expectedMessage: string,
-  consoleMessages: string[]
+  consoleMessages: string[],
+  intervals?: number[],
+  timeout?: number
 ) => {
   if (expectedMessage === undefined){
     throw new Error("expected message is null")
   }
+
+  if (intervals === undefined) {
+    intervals =  [5_000, 10_000, 15_000]
+  }
+
+  if (timeout === undefined) {
+    timeout =  60_000
+  }
+
   await expect(async () => {
     expect(messageExists(consoleMessages, expectedMessage)).toBe(true);
   }).toPass({
-    intervals: [5_000, 10_000, 15_000, 60_000, 70_000, 75_000, 85_000, 95_000],
-    timeout: 120_000,
+    intervals: intervals,
+    timeout: timeout,
   });
   
 };
